requirements.txt
streamlit>=1.25.0
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
pandas>=1.5.0
pillow>=9.0.0
plotly>=5.0.0
pathlib